+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.10 - Unreleased
 
+### Added
+* Added `skipna` argument to `kwarray.Sticher`
+
+### Changed
+* Allow `None` as `indices` in `kwarray.Sticher.add`
+
 
 ## Version 0.6.9 - Released 2023-02-02
 
@@ -19,12 +25,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.8 - Released 2023-01-11
 
-
-## Version 0.6.7 - Released 2023-01-02
-
 ### Fixed:
 * Removed extraneous print statement in `robust_normalize`
 
+## Version 0.6.7 - Released 2023-01-02
+
+### Changed:
+* Nothing, this was an accidental release.
 
 ## Version 0.6.6 - Released 2023-01-02